fix: rename getjwksurl file and function name
